Renamed ChessServer\Command\Game\Sync as ChessServer\Command\Game\Async
