+++ src/Command/Game/Async/AcceptPlayRequestCommand.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace ChessServer\Command\Game\Sync;
+namespace ChessServer\Command\Game\Async
+;
 
 use Chess\Variant\Classical\PGN\AN\Color;
 use ChessServer\Command\AbstractAsyncCommand;
