+++ src/Command/Game/Async/EvalNamesCommand.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace ChessServer\Command\Game\Sync;
+namespace ChessServer\Command\Game\Async
+;
 
 use Chess\Function\FastFunction;
 use ChessServer\Command\AbstractAsyncCommand;
