+++ src/Command/Game/Async/StartCommand.php
@@ -1,6 +1,7 @@
 <?php
 
-namespace ChessServer\Command\Game\Sync;
+namespace ChessServer\Command\Game\Async
+;
 
 use Chess\FenToBoardFactory;
 use Chess\Play\SanPlay;
